% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_double_matrix}
\alias{as_double_matrix}
\title{Convert input to a matrix with \code{"double"} storage mode (\code{\link[base:mode]{base::storage.mode()}}).}
\usage{
as_double_matrix(x)
}
\arguments{
\item{x}{\code{data.frame()}|\code{matrix()}\cr A numerical data frame or matrix with at least 1 row and 2 columns.}
}
\value{
\code{x} is coerced to a numerical \code{matrix()}.
}
\description{
Convert input to a matrix with \code{"double"} storage mode (\code{\link[base:mode]{base::storage.mode()}}).
}
