% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eafdiff.R
\name{eafdiff}
\alias{eafdiff}
\title{Compute empirical attainment function differences}
\usage{
eafdiff(x, y, intervals = NULL, maximise = FALSE, rectangles = FALSE)
}
\arguments{
\item{x, y}{\code{matrix}|\code{data.frame()}\cr Data frames corresponding to the input data of
left and right sides, respectively. Each data frame has at least three
columns, the last one is the set of each point. See also
\code{\link[=read_datasets]{read_datasets()}}.}

\item{intervals}{\code{integer(1)}\cr The absolute range of the differences
\eqn{[0, 1]} is partitioned into the number of intervals provided.}

\item{maximise}{\code{logical()}\cr Whether the objectives must be maximised
instead of minimised. Either a single logical value that applies to all
objectives or a vector of logical values, with one value per objective.}

\item{rectangles}{\code{logical(1)}\cr If TRUE, the output is in the form of rectangles of the same color.}
}
\value{
With \code{rectangle=FALSE}, a \code{data.frame} containing points where there
is a transition in the value of the EAF differences.  With
\code{rectangle=TRUE}, a \code{matrix} where the first 4 columns give the
coordinates of two corners of each rectangle. In both cases, the last
column gives the difference in terms of sets in \code{x} minus sets in \code{y} that
attain each point (i.e., negative values are differences in favour \code{y}).
}
\description{
Calculate the differences between the empirical attainment functions of two
data sets.
}
\details{
This function calculates the differences between the EAFs of two
data sets.
}
\seealso{
\code{\link[=read_datasets]{read_datasets()}}
}
\concept{eaf}
\examples{
A1 <- read_datasets(text='
 3 2
 2 3

 2.5 1
 1 2

 1 2
')

A2 <- read_datasets(text='
 4 2.5
 3 3
 2.5 3.5

 3 3
 2.5 3.5

 2 1
')
d <- eafdiff(A1, A2)
str(d)









d

d <- eafdiff(A1, A2, rectangles = TRUE)
str(d)
d
}
