% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv_approx.R
\name{hv_approx_fpras}
\alias{hv_approx_fpras}
\title{Approximate the hypervolume indicator via a fully polynomial-time randomized approximation scheme (FPRAS).}
\usage{
hv_approx_fpras(
  x,
  reference,
  maximise = FALSE,
  seed = NULL,
  epsilon = 0.01,
  delta = 0.1
)
}
\arguments{
\item{x}{\code{matrix()}|\code{data.frame()}\cr Matrix or data frame of numerical
values, where each row gives the coordinates of a point.}

\item{reference}{\code{numeric()}\cr Reference point as a vector of numerical
values.}

\item{maximise}{\code{logical()}\cr Whether the objectives must be maximised
instead of minimised. Either a single logical value that applies to all
objectives or a vector of logical values, with one value per objective.}

\item{seed}{\code{integer(1)}\cr Random seed.}

\item{epsilon}{\code{double(1)}\cr Desired relative error of the approximation,
\eqn{\epsilon > 0}.}

\item{delta}{\code{double(1)}\cr Desired failure probability
\eqn{0 < \delta < 1}; \eqn{(1 - \delta)} gives the confidence level.}
}
\value{
A single numerical value.
}
\description{
This function implements the approximation algorithm by
\citet{BriFri2010approx}.  This algorithm returns, with probability
\eqn{(1 - \delta)}, an \eqn{\epsilon}-approximation of the hypervolume
metric with respect to the given reference point, assuming minimization of
all objectives by default.
}
\details{
This function computes an approximation \eqn{\hat{v}} of the true
hypervolume \eqn{v = \text{hyp}_r(A)} of the input points in \eqn{A
\subset \mathbb{R}^m} with respect to the reference point \eqn{r \in
\mathbb{R}^m}, such that

\deqn{\text{Pr}[(1-\epsilon)v \leq \hat{v} \leq (1+\epsilon)v] \geq (1 - \delta)}

where \eqn{\epsilon > 0} and \eqn{0 < \delta < 1}.

The algorithm requires \eqn{O(\frac{nm}{\epsilon^2}\log\frac{1}{\delta})}.
That is, it is linear on the number of points and dimensions, but quadratic
in the approximation error.  In other words, more accurate approximations
require significantly more time.

In contrast to the (quasi)-Monte-Carlo methods provided by \code{\link[=hv_approx]{hv_approx()}},
the presence of weakly-dominated points not only increases the runtime, but
also changes the returned approximation for a fixed random seed.

The implementation uses Walker-Vose's alias method for sampling from a
discrete probability distribution \citep{Vose1991alias}, which requires
\eqn{O(1)} per sample.  Using the naive roulette-wheel method would add, at
least, a factor of \eqn{O(\log n)} to the above runtime.
}
\section{Warning}{

Lower values of \code{epsilon} (\eqn{\epsilon}) or \code{delta}
(\eqn{\delta}) require significantly longer computation time.
}

\examples{
x <- matrix(c(5, 5, 4, 6, 2, 7, 7, 4), ncol=2, byrow=TRUE)
hypervolume(x, ref=10)
hv_approx(x, ref=10, method="Rphi-FWE+")
hv_approx_fpras(x, ref=10, epsilon=0.1, delta=0.2, seed=42)
hv_approx_fpras(x, ref=10, epsilon=0.01, delta=0.2, seed=42)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=hypervolume]{hypervolume()}}, \code{\link[=whv_hype]{whv_hype()}}, \code{\link[=hv_approx]{hv_approx()}}
}
\author{
Manuel \enc{López-Ibáñez}{Lopez-Ibanez}
}
\concept{metrics}
