% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moocore-package.R
\docType{package}
\name{moocore-package}
\alias{moocore}
\alias{moocore-package}
\title{moocore: Core Mathematical Functions for Multi-Objective Optimization}
\description{
Fast implementation of mathematical operations and performance metrics for multi-objective optimization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://multi-objective.github.io/moocore/r/}
  \item \url{https://github.com/multi-objective/moocore/tree/main/r}
  \item Report bugs at \url{https://github.com/multi-objective/moocore/issues}
}

}
\keyword{internal}
