% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nondominated.R
\name{pareto_rank}
\alias{pareto_rank}
\title{Rank points according to Pareto-optimality (nondominated sorting).}
\usage{
pareto_rank(x, maximise = FALSE)
}
\arguments{
\item{x}{\code{matrix()}|\code{data.frame()}\cr Matrix or data frame of numerical
values, where each row gives the coordinates of a point.}

\item{maximise}{\code{logical()}\cr Whether the objectives must be maximised
instead of minimised. Either a single logical value that applies to all
objectives or a vector of logical values, with one value per objective.}
}
\value{
An integer vector of the same length as the number of rows of the
input \code{x}, where each value gives the rank of each point (lower is
better).
}
\description{
\code{\link[=pareto_rank]{pareto_rank()}} is meant to be used like \code{rank()}, but it assigns ranks
according to Pareto dominance, where rank 1 indicates those solutions not
dominated by any other solution in the input set.  Duplicated points are
assigned the same rank.  The resulting ranking can be used to partition
points into a list of matrices, each matrix representing a nondominated
front \citep{Deb02nsga2} (see examples below).
}
\details{
Given a finite set of points \eqn{X \subset \mathbb{R}^m} and \eqn{x,y \in
X}, let \eqn{x \prec y} denote that \eqn{x} dominates \eqn{y} according to
Pareto optimality. Nondominated sorting partitions the set \eqn{X} into an
ordered sequence of fronts, \eqn{F_0, F_1, \dots, F_k}, where \eqn{0 \leq k
\leq |X|}, satisfying the following conditions:

\enumerate{
\item All points are allocated to a front:
\deqn{\bigcup_{i=0}^{k} F_i = X}

\item Each point is allocated to only one front:
\deqn{F_i \cap F_j = \emptyset,\; \forall i,j \in \{0,1,\dots,k\},\, i \neq j}

\item Fronts are mutually nondominated:
\deqn{\nexists x,y \in F_i,\; x \prec y,\; \forall i=0,1,\dots,k}

\item The first front contains points not dominated by any other point:
\deqn{\forall y \in F_0,\; \nexists x \in X,\; x \prec y}

\item Every point in front \eqn{i} is dominated by at least one point in front
\eqn{i-1}:
\deqn{\forall y \in F_i,\; \exists x \in F_{i-1},\; x \prec y,\; \forall i=1,2,\dots,k}
}

The rank returned by \code{\link[=pareto_rank]{pareto_rank()}} is the index \eqn{i \in
\{0,1,\dots,k\}} of the front \eqn{F_i} allocated to each point. If all
points are mutually nondominated, there is only one front (\eqn{k = 0}). If
each front contains one point, then \eqn{k = n = |X|}.

With \eqn{m=2}, i.e., \code{ncol(data)=2}, the code uses the best-known \eqn{O(n
\log n)} algorithm by \citet{Jen03}.  When \eqn{m \geq 3}, it uses the naive
algorithm that identifies one \emph{front} at a time, which requires
\eqn{O(n^2\log n)} for \eqn{m=3}, and \eqn{O(n^2 \log^{m-2} n)} for \eqn{m
\geq 4}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=is_nondominated]{is_nondominated()}}
}
\concept{dominance}
\examples{
three_fronts = matrix(c(1, 2, 3,
                        3, 1, 2,
                        2, 3, 1,
                        10, 20, 30,
                        30, 10, 20,
                        20, 30, 10,
                        100, 200, 300,
                        300, 100, 200,
                        200, 300, 100), ncol=3, byrow=TRUE)
pareto_rank(three_fronts)

split.data.frame(three_fronts, pareto_rank(three_fronts))
path_A1 <- file.path(system.file(package="moocore"),"extdata","ALG_1_dat.xz")
set <- read_datasets(path_A1)[,1:2]
ranks <- pareto_rank(set)
str(ranks)
if (requireNamespace("graphics", quietly = TRUE)) {
   colors <- colorRampPalette(c("red","yellow","springgreen","royalblue"))(max(ranks))
   plot(set, col = colors[ranks], type = "p", pch = 20)
}
}
