% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moocore-package.R
\docType{data}
\name{SPEA2minstoptimeRichmond}
\alias{SPEA2minstoptimeRichmond}
\title{Results of SPEA2 when minimising electrical cost and maximising the
minimum idle time of pumps on Richmond water network.}
\format{
A data frame as produced by \code{\link[=read_datasets]{read_datasets()}}. The second
column measures time in seconds and corresponds to a maximisation problem.
}
\source{
\insertRef{LopezIbanezPhD}{moocore}
}
\usage{
SPEA2minstoptimeRichmond
}
\description{
Results of SPEA2 when minimising electrical cost and maximising the
minimum idle time of pumps on Richmond water network.
}
\examples{
data(SPEA2minstoptimeRichmond)
str(SPEA2minstoptimeRichmond)
}
\keyword{datasets}
