% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moocore-package.R
\docType{data}
\name{SPEA2relativeRichmond}
\alias{SPEA2relativeRichmond}
\title{Results of SPEA2 with relative time-controlled triggers on Richmond water
network.}
\format{
A data frame as produced by \code{\link[=read_datasets]{read_datasets()}}.
}
\source{
\insertRef{LopezIbanezPhD}{moocore}
}
\usage{
SPEA2relativeRichmond
}
\description{
Results of SPEA2 with relative time-controlled triggers on Richmond water
network.
}
\examples{
data(HybridGA)
data(SPEA2relativeRichmond)
}
\keyword{datasets}
