% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moocore-package.R
\docType{data}
\name{SPEA2relativeVanzyl}
\alias{SPEA2relativeVanzyl}
\title{Results of SPEA2 with relative time-controlled triggers on Vanzyl's
water network.}
\format{
A data frame as produced by \code{\link[=read_datasets]{read_datasets()}}.
}
\source{
\insertRef{LopezIbanezPhD}{moocore}
}
\usage{
SPEA2relativeVanzyl
}
\description{
Results of SPEA2 with relative time-controlled triggers on Vanzyl's
water network.
}
\examples{
data(HybridGA)
data(SPEA2relativeVanzyl)
}
\keyword{datasets}
