% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moocore-package.R
\docType{data}
\name{tpls50x20_1_MWT}
\alias{tpls50x20_1_MWT}
\title{Various strategies of Two-Phase Local Search applied to the Permutation
Flowshop Problem with Makespan and Weighted Tardiness objectives.}
\format{
A data frame with 1511 observations of 4 variables:
\describe{
\item{\code{algorithm}}{TPLS search strategy}
\item{\code{Makespan}}{first objective values.}
\item{\code{WeightedTardiness}}{second objective values.}
\item{\code{run}}{index of the run.}
}
}
\source{
\insertRef{DubLopStu2011amai}{moocore}
}
\usage{
tpls50x20_1_MWT
}
\description{
Various strategies of Two-Phase Local Search applied to the Permutation
Flowshop Problem with Makespan and Weighted Tardiness objectives.
}
\examples{
data(tpls50x20_1_MWT)
str(tpls50x20_1_MWT)
}
\keyword{datasets}
