% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datasets.R
\name{write_datasets}
\alias{write_datasets}
\title{Write data sets}
\usage{
write_datasets(x, file = "")
}
\arguments{
\item{x}{\code{matrix}|\code{data.frame()}\cr Dataset with at least three
columns, the last one is the set of each point. See also
\code{\link[=read_datasets]{read_datasets()}}.}

\item{file}{Either a character string naming a file or a connection open for
writing. \code{""} indicates output to the console.}
}
\value{
No return value, called for side effects
}
\description{
Write data sets to a file in the same format as \code{\link[=read_datasets]{read_datasets()}}.
}
\examples{
x <- read_datasets(text="1 2\n3 4\n\n5 6\n7 8\n", col_names=c("obj1", "obj2"))
write_datasets(x)

}
\seealso{
\code{\link[utils:write.table]{utils::write.table()}}, \code{\link[=read_datasets]{read_datasets()}}
}
\keyword{file}
