% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moocore-package.R
\docType{package}
\name{moocore-package}
\alias{moocore}
\alias{moocore-package}
\title{moocore: Core Mathematical Functions for Multi-Objective Optimization}
\description{
Fast implementation of mathematical operations and performance metrics for multi-objective optimization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://multi-objective.github.io/moocore/r/}
  \item \url{https://github.com/multi-objective/moocore/tree/main/r}
  \item Report bugs at \url{https://github.com/multi-objective/moocore/issues}
}

}
\author{
\strong{Maintainer}: Manuel López-Ibáñez \email{manuel.lopez-ibanez@manchester.ac.uk} (\href{https://orcid.org/0000-0001-9974-1295}{ORCID})

Authors:
\itemize{
  \item Carlos Fonseca
  \item Luís Paquete
}

Other contributors:
\itemize{
  \item Mickaël Binois [contributor]
}

}
\keyword{internal}
