% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moocore-package.R
\docType{package}
\name{moocore-package}
\alias{moocore}
\alias{moocore-package}
\title{moocore: Core Mathematical Functions for Multi-Objective Optimization}
\description{
Fast implementation of mathematical operations and performance metrics for multi-objective optimization, including filtering and ranking of dominated vectors according to Pareto optimality, computation of the empirical attainment function, V.G. da Fonseca, C.M. Fonseca, A.O. Hall (2001) \doi{10.1007/3-540-44719-9_15}, hypervolume metric, C.M. Fonseca, L. Paquete, M. López-Ibáñez (2006) \doi{10.1109/CEC.2006.1688440}, epsilon indicator, inverted generational distance, and Vorob'ev threshold, expectation and deviation, M. Binois, D. Ginsbourger, O. Roustant (2015) \doi{10.1016/j.ejor.2014.07.032}, among others.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://multi-objective.github.io/moocore/r/}
  \item \url{https://github.com/multi-objective/moocore/}
  \item Report bugs at \url{https://github.com/multi-objective/moocore/issues}
}

}
\author{
\strong{Maintainer}: Manuel López-Ibáñez \email{manuel.lopez-ibanez@manchester.ac.uk} (\href{https://orcid.org/0000-0001-9974-1295}{ORCID})

Other contributors:
\itemize{
  \item Carlos Fonseca [contributor]
  \item Luís Paquete [contributor]
  \item Andreia P. Guerreiro [contributor]
  \item Mickaël Binois [contributor]
  \item Michael H. Buselli (AVL-tree library) [copyright holder]
  \item Wessel Dankers (AVL-tree library) [copyright holder]
  \item NumPy Developers (RNG and ziggurat constants) [copyright holder]
  \item Jean-Sebastien Roy (mt19937 library) [copyright holder]
  \item Makoto Matsumoto (mt19937 library) [copyright holder]
  \item Takuji Nishimura (mt19937 library) [copyright holder]
}

}
\keyword{internal}
