% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moocore-package.R
\docType{package}
\name{moocore-package}
\alias{moocore}
\alias{moocore-package}
\title{SHORT DESCRIPTION}
\description{
LONG DESCRIPTION
}
\examples{
# TODO
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/multi-objective/moocore/,}
  \item Report bugs at \url{https://github.com/multi-objective/moocore/issues}
}

}
\author{
\strong{Maintainer}: Manuel López-Ibáñez \email{manuel.lopez-ibanez@manchester.ac.uk} (\href{https://orcid.org/0000-0001-9974-1295}{ORCID})

}
\concept{empirical attainment function}
\concept{multivariate}
\concept{optimize}
\keyword{internal}
