% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2_exact}
\alias{r2_exact}
\title{Exact R2 indicator}
\usage{
r2_exact(x, reference, maximise = FALSE)
}
\arguments{
\item{x}{\code{matrix()}|\code{data.frame()}\cr Matrix or data frame of numerical
values, where each row gives the coordinates of a point.}

\item{reference}{\code{numeric()}\cr Reference point as a vector of numerical
values.}

\item{maximise}{\code{logical()}\cr Whether the objectives must be maximised
instead of minimised. Either a single logical value that applies to all
objectives or a vector of logical values, with one value per objective.}
}
\value{
\code{numeric(1)} A single numerical value.
}
\description{
Computes the exact R2 indicator with respect to a given ideal/utopian
reference point assuming minimization of all objectives.
}
\details{
The current implementation exclusively supports bi-objective solution sets
and runs in \eqn{O(n \log n)}.  For more details on the computation of the
exact R2 indicator refer to \citet{SchKer2025r2v2}.
}
\examples{
dat <- matrix(c(5, 5, 4, 6, 2, 7, 7, 4), ncol = 2, byrow = TRUE)
r2_exact(dat, reference = c(0, 0))

# This function assumes minimisation by default. We can easily specify maximisation:
r2_exact(dat, reference = c(10, 10), maximise = TRUE)

# Merge all the sets of a dataset by removing the set number column:
extdata_path <- system.file(package="moocore","extdata")
dat <- read_datasets(file.path(extdata_path, "example1_dat"))[, 1:2]
nrow(dat)

# Dominated points are ignored, so this:
r2_exact(dat, reference = 0)

# gives the same exact R2 value as this:
dat <- filter_dominated(dat)
nrow(dat)
r2_exact(dat, reference = 0)

}
\references{
\insertAllCited{}
}
\author{
Lennart \enc{Schäpermeier}{Schapermeier}
}
\concept{metrics}
