% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_datasets.R
\name{rbind_datasets}
\alias{rbind_datasets}
\title{Combine datasets \code{x} and \code{y} by row taking care of making all sets unique.}
\usage{
rbind_datasets(x, y)
}
\arguments{
\item{x, y}{Datasets.}
}
\value{
A dataset.
}
\description{
Combine datasets \code{x} and \code{y} by row taking care of making all sets unique.
}
