% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moocore-package.R
\docType{data}
\name{SPEA2relativeVanzyl}
\alias{SPEA2relativeVanzyl}
\title{Results of SPEA2 with relative time-controlled triggers on Vanzyl's
water network.}
\format{
An object of class \code{data.frame} with 107 rows and 3 columns.
}
\source{
\insertRef{LopezIbanezPhD}{moocore}
}
\usage{
SPEA2relativeVanzyl
}
\description{
Results of SPEA2 with relative time-controlled triggers on Vanzyl's
water network.
}
\examples{
data(SPEA2relativeVanzyl)
str(SPEA2relativeVanzyl)
}
\keyword{datasets}
